pyproject.toml
[tool.poetry]
name = "storytime"
version = "0.1.0"
description = "An application for generating illustrations to using generative AI based on the spoken word."
authors = ["Thomas Lagrange"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.listening_ears.dependencies]
openai-whisper = {git = "https://github.com/openai/whisper.git"}
pyaudio = "^0.2.14"
ffmpeg-python = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
